refactor(components): folder structure
no sub folders
