+++ src/components/ToClipboard.js
